pyproject.toml
[project]
name = "satellite-contest"
version = "0.1.0"
description = "Color sorting and Telegram gift collections visualization"
requires-python = ">=3.10"
dependencies = [
    "pillow>=10.0.0",
]

[tool.uv]
dev-dependencies = []
